refactor: change f64 to f32 for memory optimization and add rounding
Convert all floating-point numbers from f64 to f32 to reduce memory usage and improve performance. Add rounding to 3 decimal places when converting to JS values to prevent floating-point precision issues.

+++ rust/src/libs/convas_optimization/canvas_optimization.rs
@@ -27,10 +27,10 @@ pub struct ShapeData {
 
 #[derive(Serialize)]
 pub struct AsValues {
-    pub as1: Vec<f64>,
-    pub as2: Vec<f64>,
-    pub as3: Vec<f64>,
-    pub as4: Vec<f64>,
+    pub as1: Vec<f32>,
+    pub as2: Vec<f32>,
+    pub as3: Vec<f32>,
+    pub as4: Vec<f32>,
 }
 
 #[derive(Serialize)]
